backend/requirements.txt
fastapi
sqlalchemy
uvicorn
psycopg2-binary
python-dotenv
pydantic
celery
redis
azure-storage-blob
azure-ai-formrecognizer
python-multipart
alembic
openpyxl
aiohttp
openai